+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Minor typo
 - You can no longer navigate to other websites from the release description
+- Driver installer no longer hardcoded to run quiet if minimal installer is used
 
 ## [1.13.0] - 2018-06-06
 ### Added